+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Back to development.
+- Added `fx.Optional` embedded type to mark parameter structs as optional.
 
 ## [1.10.0] - 2019-11-20
 ### Added